# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kubernetes REST API",
    "version": "1.0.0",
    "description": "A simplified REST API for managing Kubernetes resources using client-go. Supports pods, deployments, services, namespaces, configmaps, persistent volume claims, nodes, and cluster info."
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
  "paths": {
...
    "/namespaces": {
      "get": {
        "summary": "List namespaces",
        "operationId": "listNamespaces",
        "tags": ["Namespaces"],
        "responses": {
          "200": {
            "description": "List of namespace summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/NamespaceSummary" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create a namespace",
        "operationId": "createNamespace",
        "tags": ["Namespaces"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateNamespaceRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Namespace created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NamespaceSummary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/namespaces/{name}": {
      "get": {
        "summary": "Inspect a namespace",
        "operationId": "getNamespace",
        "tags": ["Namespaces"],
        "parameters": [
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "200": {
            "description": "Namespace object",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NamespaceSummary" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete a namespace",
        "description": "Deletes a namespace and all resources within it.",
        "operationId": "deleteNamespace",
        "tags": ["Namespaces"],
        "parameters": [
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "204": { "description": "Namespace deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/nodes": {
      "get": {
        "summary": "List nodes",
        "description": "Returns all nodes in the cluster with capacity, allocatable resources, and conditions.",
        "operationId": "listNodes",
        "tags": ["Nodes"],
        "responses": {
          "200": {
            "description": "List of node summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/NodeSummary" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/nodes/{name}": {
      "get": {
        "summary": "Inspect a node",
        "operationId": "getNode",
        "tags": ["Nodes"],
        "parameters": [
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "200": {
            "description": "Node summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NodeSummary" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "name": {
        "name": "name",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Resource name"
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or missing required fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InternalError": {
        "description": "Kubernetes API error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "pod not found" }
        }
      },
...
      "NamespaceSummary": {
        "type": "object",
        "properties": {
          "name":       { "type": "string" },
          "status":     { "type": "string", "example": "Active" },
          "labels":     { "type": "object", "additionalProperties": { "type": "string" } },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateNamespaceRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name":   { "type": "string" },
          "labels": { "type": "object", "additionalProperties": { "type": "string" } }
        }
      },
...
      "NodeCondition": {
        "type": "object",
        "properties": {
          "type":    { "type": "string" },
          "status":  { "type": "string" },
          "message": { "type": "string" },
          "reason":  { "type": "string" }
        }
      },
      "NodeAddress": {
        "type": "object",
        "properties": {
          "type":    { "type": "string", "example": "InternalIP" },
          "address": { "type": "string", "example": "10.0.0.1" }
        }
      },
      "NodeSummary": {
        "type": "object",
        "properties": {
          "name":               { "type": "string" },
          "status":             { "type": "string", "example": "Ready" },
          "roles":              { "type": "array", "items": { "type": "string" } },
          "kubelet_version":    { "type": "string" },
          "os_image":           { "type": "string" },
          "container_runtime":  { "type": "string" },
          "architecture":       { "type": "string" },
          "cpu_capacity":       { "type": "string" },
          "memory_capacity":    { "type": "string" },
          "cpu_allocatable":    { "type": "string" },
          "memory_allocatable": { "type": "string" },
          "labels":             { "type": "object", "additionalProperties": { "type": "string" } },
          "conditions":         { "type": "array", "items": { "$ref": "#/components/schemas/NodeCondition" } },
          "addresses":          { "type": "array", "items": { "$ref": "#/components/schemas/NodeAddress" } },
          "created_at":         { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
}